+++ packages/core/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## [2.0.1] - 2025-11-12
+
+- Remove `.xml` and `.svg` file type support
+
+## [2.0.0] - 2025-11-05
+
+- Updates valueMatchesType to be more inline with `isFieldResponseValid` returns a t(string).
+
+- Adds valueMatches return <Boolean> use in place of `isFieldResponseValid` where a <Boolean> is needed.
+
+## [1.0.11] - 2025-10-29
+
+- Remove `.numbers` file type support
+
+## [1.0.10] - 2025-10-28
+
+- Add values-to-types and isValidEmail
+
+## [1.0.9] - 2025-10-27
+
+- Add exports field to package.json
+
 ## [1.0.8] - 2025-10-01
 
 - Bump yarn